Export common types
