+++ packages/api/src/beacon/client/node.ts
@@ -1,6 +1,5 @@
 import {IChainForkConfig} from "@lodestar/config";
 import {ClientOptions, generateGenericJsonClient, IHttpClient} from "../../utils/client/index.js";
-import {ServerApi} from "../../utils/types.js";
 import {Api, getReqSerializers, getReturnTypes, ReqTypes, routesData} from "../routes/node.js";
 import {defaultClientOptions} from "./index.js";
