+++ packages/api/src/beacon/routes/beacon/index.ts
@@ -1,7 +1,8 @@
 import {IChainForkConfig} from "@lodestar/config";
 import {phase0, ssz} from "@lodestar/types";
+import {ApiClientResponse} from "../../../interfaces.js";
 import {HttpStatusCode} from "../../../utils/client/httpStatusCode.js";
-import {RoutesData, ReturnTypes, reqEmpty, ContainerData, ApiClientResponse} from "../../../utils/index.js";
+import {RoutesData, ReturnTypes, reqEmpty, ContainerData} from "../../../utils/index.js";
 import * as block from "./block.js";
 import * as pool from "./pool.js";
 import * as state from "./state.js";
