+++ packages/api/src/beacon/routes/config.ts
@@ -11,10 +11,10 @@ import {
   ReqSerializers,
   RoutesData,
   sameType,
-  ApiClientResponse,
   ContainerData,
 } from "../../utils/index.js";
 import {HttpStatusCode} from "../../utils/client/httpStatusCode.js";
+import {ApiClientResponse} from "../../interfaces.js";
 
 // See /packages/api/src/routes/index.ts for reasoning and instructions to add new routes
