+++ packages/api/src/beacon/routes/node.ts
@@ -9,10 +9,10 @@ import {
   Schema,
   ReqSerializers,
   ReqEmpty,
-  ApiClientResponse,
   ContainerData,
 } from "../../utils/index.js";
 import {HttpStatusCode} from "../../utils/client/httpStatusCode.js";
+import {ApiClientResponse} from "../../interfaces.js";
 
 // See /packages/api/src/routes/index.ts for reasoning and instructions to add new routes
