+++ packages/api/src/beacon/server/beacon.ts
@@ -1,7 +1,7 @@
 import {IChainForkConfig} from "@lodestar/config";
 import {Api, ReqTypes, routesData, getReturnTypes, getReqSerializers} from "../routes/beacon/index.js";
 import {ServerRoutes, getGenericJsonServer} from "../../utils/server/index.js";
-import {ServerApi} from "../../utils/types.js";
+import {ServerApi} from "../../interfaces.js";
 
 export function getRoutes(config: IChainForkConfig, api: ServerApi<Api>): ServerRoutes<Api, ReqTypes> {
   // All routes return JSON, use a server auto-generator
