+++ packages/api/src/index.ts
@@ -1,6 +1,7 @@
 // Re-exporting beacon only for backwards compatibility
 export * from "./beacon/index.js";
-
+export * from "./interfaces.js";
+export {HttpStatusCode, HttpErrorCodes, HttpSuccessCodes} from "./utils/client/httpStatusCode.js";
 export {HttpClient, IHttpClient, HttpClientOptions, HttpClientModules, HttpError} from "./utils/client/index.js";
 
 // NOTE: Don't export server here so it's not bundled to all consumers
